combine LifecycleLimiter and LifecycleRegistry
